+++ src/svd/field.rs
@@ -1,4 +1,4 @@
-use core::ops::Deref;
+use core::ops::{Deref, DerefMut};
 
 use xmltree::Element;
 
@@ -21,6 +21,15 @@ impl Deref for Field {
     type Target = FieldInfo;
 
     fn deref(&self) -> &FieldInfo {
+        match self {
+            Field::Single(info) => info,
+            Field::Array(info, _) => info,
+        }
+    }
+}
+
+impl DerefMut for Field {
+    fn deref_mut(&mut self) -> &mut FieldInfo {
         match self {
             Field::Single(info) => info,
             Field::Array(info, _) => info,